requirements.txt
flask-security-too[babel, fsqla, common, mfa]
flask_restful
flask_cors
celery
redis